requirements.txt
flask
gunicorn
python-dotenv

sentence-transformers
scikit-learn

openai>=1.0.0
